Feat: Add test for getAllAlbumsByReleaseYear in AlbumManagerController when there are matching albums DB for given year and implementation
Test: added test for getAllAlbumsByReleaseYear method in AlbumManagerController.

Implemented logic in getAllAlbumsByReleaseYear in AlbumManagerController.

+++ src/main/java/com/northcoders/recordshop/service/AlbumManagerService.java
@@ -13,7 +13,7 @@ public interface AlbumManagerService {
 
     List<Album> getAllAlbumsByArtist(String artistName);
 
-    List<Album> getAllAlbumsByReleaseYear(Long year);
+    List<Album> getAllAlbumsByReleaseYear(int year);
 
     List<Album> getAllAlbumsByGenre(Genre genre);
